+++ libs/deepagents/src/middleware/subagent.test.ts
@@ -835,51 +835,3 @@ describe("ls_agent_type tracing metadata", () => {
     expect(subagentRuns.length).toBeGreaterThan(0);
   });
 });
-
-/**
- * Tests for task tool subagent type validation.
- */
-describe("task tool subagent validation", () => {
-  it("throws a descriptive error for an unknown subagent_type", async () => {
-    const model = new FakeListChatModel({
-      responses: [
-        new AIMessage({
-          content: "",
-          tool_calls: [
-            {
-              id: `call_${Date.now()}`,
-              name: "task",
-              args: {
-                description: "Work",
-                subagent_type: "ghost-agent",
-              },
-            },
-          ],
-        }) as unknown as string,
-        "Done",
-      ],
-    });
-
-    const agent = createDeepAgent({
-      model,
-      checkpointer: new MemorySaver(),
-      subagents: [
-        {
-          name: "analyst",
-          description: "An analyst",
-          systemPrompt: "You are an analyst.",
-        },
-      ],
-    });
-
-    await expect(
-      agent.invoke(
-        { messages: [new HumanMessage("Test")] },
-        {
-          configurable: { thread_id: `unknown-type-${Date.now()}` },
-          recursionLimit: 50,
-        },
-      ),
-    ).rejects.toThrow("ghost-agent");
-  });
-});
